kill a process

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,4 +1,7 @@
 # Define an exec resource to kill the process
 exec { 'killmenow':
-  command     => 'pkill -f killmenow',
+  command => 'pkill -f killmenow',
+  path    => ['/usr/bin', '/bin'],
+  onlyif  => 'pgrep -f killmenow', # Check if the process is running
 }
+
